Renamed rc to web RC and made changes in all files linked to it
