+++ core/templates/pages/web-release-coordinator-page/features-tab/features-tab.component.ts
@@ -38,7 +38,7 @@ import {FeatureFlagBackendApiService} from 'domain/feature-flag/feature-flag-bac
 import {PlatformFeatureService} from 'services/platform-feature.service';
 import {FeatureFlag} from 'domain/feature-flag/feature-flag.model';
 import {HttpErrorResponse} from '@angular/common/http';
-import {UserGroup} from 'domain/release_coordinator/user-group.model';
+import {UserGroup} from 'domain/web_release_coordinator/user-group.model';
 
 interface IntSchema {
   type: 'int';
